pvr.dvblink version 2.2.0 (jarvis) Fixed: Crash on startup with search timers Fixed: Incorrect disk space display if disk size is more than 2TB Fixed: Doesn't build with tinyxml2 4.x Added: transcoded recordings playback in DVBLink v6

+++ lib/libdvblinkremote/response.h
@@ -1350,12 +1350,12 @@ namespace dvblinkremote {
     /**
       * The total space in KB.
       */
-    long TotalSpace;
+    long long TotalSpace;
 
     /**
       * The available space in KB. 
       */
-    long AvailableSpace;
+    long long AvailableSpace;
   };
 
   class ChannelFavorite
